requirements.txt
aiohttp==3.6.3
async-timeout==3.0.1
attrs==20.3.0
certifi==2020.12.5
chardet==3.0.4
discord.py==1.5.1
idna==2.10
multidict==4.7.6
python-dateutil==2.8.1
six==1.15.0
text-unidecode==1.3
urllib3==1.26.4
yarl==1.5.1
